add useModal hook, upgrade drawer
