+++ src/shared/ui/Drawer/Drawer.tsx
@@ -5,6 +5,7 @@ import {
     classNames,
     Mods,
 } from 'shared/lib/classNames/classNames';
+import { useModal } from 'shared/lib/hooks/useModal/useModal';
 
 import { Overlay } from '../Overlay/Overlay';
 import { Portal } from '../Portal/Portal';
@@ -16,6 +17,7 @@ interface DrawerProps {
     children: ReactNode;
     isOpen?: boolean;
     onClose?: () => void;
+    lazy?: boolean;
 }
 
 export const Drawer = memo((props: DrawerProps) => {
@@ -24,17 +26,28 @@ export const Drawer = memo((props: DrawerProps) => {
         children,
         onClose,
         isOpen,
+        lazy,
     } = props;
     const { theme } = useTheme();
 
+    const { isClosing, close, isMounted } = useModal({
+        animationDelay: 300,
+        onClose,
+        isOpen,
+    });
+
     const mods: Mods = {
         [cls.opened]: isOpen,
+        [cls.isClosing]: isClosing,
     };
 
+    if (lazy && !isMounted) {
+        return null;
+    }
     return (
         <Portal>
             <div className={classNames(cls.Drawer, mods, [className, theme, 'app_drawer'])}>
-                <Overlay onClick={onClose} />
+                <Overlay onClick={close} />
                 <div
                     className={cls.content}
                 >
